requirements.txt
# Python Libraries
pandas
numpy
# Database Connector
psycopg2-binary 
sqlalchemy
# Data Visualization / Notebook
matplotlib
jupyter
# Environment Variables
python-dotenv 
# For API Integration (Próxima etapa!)
requests